fix: imports for renamed utils file
